git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp-misc/trunk/slideshow@128 acdd8558-d4fd-0310-b0a5-8cf3c1ec824a

+++ src/converter.cc
@@ -57,17 +57,17 @@ void scaleImageSize(size_t  originalWidth,
 
 
 // ###### Convert image to fullsize and preview versions ####################
-int imageConverter(const char*   originalImageName,
-                   const char*   previewImageName,
-                   const char*   fullsizeImageName,
-                   unsigned int& originalWidth,
-                   unsigned int& originalHeight,
-                   unsigned int& previewWidth,
-                   unsigned int& previewHeight,
-                   unsigned int  previewQuality,
-                   unsigned int& fullsizeWidth,
-                   unsigned int& fullsizeHeight,
-                   unsigned int  fullsizeQuality)
+int imageConverter(const char* originalImageName,
+                   const char* previewImageName,
+                   const char* fullsizeImageName,
+                   size_t&     originalWidth,
+                   size_t&     originalHeight,
+                   size_t&     previewWidth,
+                   size_t&     previewHeight,
+                   size_t      previewQuality,
+                   size_t&     fullsizeWidth,
+                   size_t&     fullsizeHeight,
+                   size_t      fullsizeQuality)
 {
    ExceptionInfo exception;
    ImageInfo*    originalImageInfo;
@@ -161,15 +161,15 @@ int imageConverter(const char*   originalImageName,
 
 
 // ###### Test original, fullsize and preview images ########################
-int imageTester(const char*   originalImageName,
-                const char*   previewImageName,
-                const char*   fullsizeImageName,
-                unsigned int& originalWidth,
-                unsigned int& originalHeight,
-                unsigned int& previewWidth,
-                unsigned int& previewHeight,
-                unsigned int& fullsizeWidth,
-                unsigned int& fullsizeHeight)
+int imageTester(const char* originalImageName,
+                const char* previewImageName,
+                const char* fullsizeImageName,
+                size_t&     originalWidth,
+                size_t&     originalHeight,
+                size_t&     previewWidth,
+                size_t&     previewHeight,
+                size_t&     fullsizeWidth,
+                size_t&     fullsizeHeight)
 {
    ExceptionInfo exception;
    ImageInfo*    originalImageInfo;
