+++ src/converter.h
@@ -17,27 +17,27 @@
 #define CONVERTER_H
 
 
-int imageConverter(const char*   originalImageName,
-                   const char*   previewImageName,
-                   const char*   fullsizeImageName,
-                   unsigned int& originalWidth,
-                   unsigned int& originalHeight,
-                   unsigned int& previewWidth,
-                   unsigned int& previewHeight,
-                   unsigned int  previewQuality,
-                   unsigned int& fullsizeWidth,
-                   unsigned int& fullsizeHeight,
-                   unsigned int  fullsizeQuality);
+int imageConverter(const char* originalImageName,
+                   const char* previewImageName,
+                   const char* fullsizeImageName,
+                   size_t&     originalWidth,
+                   size_t&     originalHeight,
+                   size_t&     previewWidth,
+                   size_t&     previewHeight,
+                   size_t      previewQuality,
+                   size_t&     fullsizeWidth,
+                   size_t&     fullsizeHeight,
+                   size_t      fullsizeQuality);
 
-int imageTester(const char*   originalImageName,
-                const char*   previewImageName,
-                const char*   fullsizeImageName,
-                unsigned int& originalWidth,
-                unsigned int& originalHeight,
-                unsigned int& previewWidth,
-                unsigned int& previewHeight,
-                unsigned int& fullsizeWidth,
-                unsigned int& fullsizeHeight);
+int imageTester(const char* originalImageName,
+                const char* previewImageName,
+                const char* fullsizeImageName,
+                size_t&     originalWidth,
+                size_t&     originalHeight,
+                size_t&     previewWidth,
+                size_t&     previewHeight,
+                size_t&     fullsizeWidth,
+                size_t&     fullsizeHeight);
 
 
 #endif
